requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
google-generativeai==0.3.2
python-dotenv==1.0.1
scikit-learn==1.4.0
sentence-transformers==2.5.1
gunicorn==21.2.0
fastapi==0.109.2
uvicorn==0.27.1

